# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: EDU Power Push Backend API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    API for campus electricity scanning, product user accounts, personal and
    campus statistics, push notifications, and the operator admin panel.
    Processes bind to loopback by default (Compose maps only 127.0.0.1). Public
    access is same-origin reverse proxy + TLS; operators may also use an SSH
    tunnel. Campus aggregates (`/api/v1/campus/*`) are anonymous; personal data
    uses short-lived access JWTs; admin and raw meter endpoints require an
    operator/admin session or `ADMIN_TOKEN` (never ship that token in a browser build).
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/admin/settings/captcha:
    get:
      tags: [Admin]
      operationId: getCaptchaSettings
      summary: Get the effective human-verification provider and protected actions
      security: [{ userAccessToken: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Effective configuration with masked secret state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptchaSettingsView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Admin]
      operationId: putCaptchaSettings
      summary: Save the provider and action-level human-verification policy
      description: >
        Secret fields are three-state: omit to keep the stored value, send null or an empty
        string to clear it, or send a new string to replace it. Provider changes take effect
        without restarting the API. Stored credentials require SETTINGS_ENCRYPTION_KEY.
      security: [{ userAccessToken: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              additionalProperties: false
              properties:
                settings:
                  $ref: "#/components/schemas/CaptchaSettings"
                secrets:
                  type: object
                  properties:
                    turnstile_secret: { type: [string, "null"] }
                  additionalProperties: false
      responses:
        "200":
          description: Saved configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptchaSettingsView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "412":
          description: Server has no SETTINGS_ENCRYPTION_KEY, so credentials cannot be stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Internal administrator token from ADMIN_TOKEN.
    userAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived product-user access JWT.
    refreshCookie:
      type: apiKey
      in: cookie
      name: edu_power_refresh
      description: Long-lived rotating refresh JWT stored as an HttpOnly SameSite cookie.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid administrator token, access JWT, or refresh JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated operation is disabled by product policy
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    CaptchaProvider:
      type: string
      enum: [disabled, turnstile]
    CaptchaAction:
      type: string
      enum: [auth.login, auth.register_code, auth.password_reset_code, meter.preview]
    CaptchaSettings:
      type: object
      additionalProperties: false
      required: [provider, site_key, hostname, actions]
      properties:
        provider:
          $ref: "#/components/schemas/CaptchaProvider"
        site_key:
          type: string
          description: Public Turnstile widget key; unused when verification is disabled.
        hostname:
          type: string
          description: Optional hostname required in a successful Turnstile verification response.
        actions:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/CaptchaAction"
...
    CaptchaSettingsView:
      type: object
      required:
        [settings, source, providers, protected_actions, secret_set, secret_masked,
         secrets_writable, secrets_unreadable, missing, ready, updated_at, updated_by]
      properties:
        settings:
          $ref: "#/components/schemas/CaptchaSettings"
        source: { type: string, enum: [env, panel] }
        providers:
          type: array
          items:
            $ref: "#/components/schemas/CaptchaProvider"
        protected_actions:
          type: array
          items:
            $ref: "#/components/schemas/CaptchaAction"
        secret_set:
          type: object
          additionalProperties: { type: boolean }
        secret_masked:
          type: object
          additionalProperties: { type: string }
        secrets_writable: { type: boolean }
        secrets_unreadable: { type: boolean }
        missing:
          type: array
          items: { type: string }
        ready: { type: boolean }
        updated_at: { type: [string, "null"], format: date-time }
        updated_by: { type: string }
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        provider:
          $ref: "#/components/schemas/CaptchaProvider"
        challenge_url:
          type: string
          format: uri-reference
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...